ia_microservices/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
redis==5.2.0
numpy==2.1.3
scikit-learn==1.5.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
